requirements.txt
asgiref==3.3.1
astroid==2.0.4
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
coverage==4.4.1
dj-database-url==0.4.2
Django==3.1.4
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-environ==0.4.4
django-js-asset==1.2.2
django-tables2==2.3.1
gunicorn==19.7.1
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.18.3
pandas==1.0.3
Pillow==7.0.0
psycopg2==2.8.5
pylint==1.7.2
pypng==0.0.20
PyQRCode==1.2.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.18.4
selenium==3.5.0
six==1.10.0
sqlparse==0.4.1
typed-ast==1.1.0
urllib3==1.22
virtualenv==16.7.5
virtualenvwrapper-win==1.2.5
whitenoise==4.1.4
wrapt==1.10.11
